Move actual type declarations from index.d.ts to module.d.ts.
