+++ package.json
@@ -5,7 +5,7 @@
   "license": "MIT",
   "main": "index.js",
   "module": "module.js",
-  "types": "index.d.ts",
+  "types": "module.d.ts",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/apollographql/zen-observable-ts.git"
